+++ src/Common/ScriptingGenerators/LuaComponentsGenerator.cs
@@ -153,12 +153,14 @@ public class LuaComponentsGenerator : IIncrementalGenerator
                     try 
                     {{
                         var argCount = lua_gettop(luaState);
-                        if (argCount != 1) throw new LuaException(""Must be called with one argument."");
+                        if (argCount < 1 || argCount > 2) throw new LuaException(""Must be called with one or two arguments."");
                         if (!lua_isinteger(luaState, 1)) throw new LuaException(""First argument must be integer."");
+                        if (argCount == 2 && !lua_isboolean(luaState, 2)) throw new LuaException(""Second argument must be boolean."");
 
                         var entityId = (ulong)lua_tointeger(luaState, 1);
+                        var lookback = argCount == 2 ? lua_toboolean(luaState, 2) : false;
 
-                        result = components.HasComponentForEntity(entityId);
+                        result = components.HasComponentForEntity(entityId, lookback);
                     }}
                     catch (Exception e)
                     {{
@@ -179,17 +181,19 @@ public class LuaComponentsGenerator : IIncrementalGenerator
                         luaL_checkstack(luaState, 1, null);
 
                         var argCount = lua_gettop(luaState);
-                        if (argCount != 1) throw new LuaException(""Must be called with one argument."");
+                        if (argCount < 1 || argCount > 2) throw new LuaException(""Must be called with one or two arguments."");
                         if (!lua_isinteger(luaState, 1)) throw new LuaException(""First argument must be integer."");
+                        if (argCount == 2 && !lua_isboolean(luaState, 2)) throw new LuaException(""Second argument must be boolean."");
 
-                        var entityId = (ulong)lua_tointeger(luaState, -1);");
+                        var entityId = (ulong)lua_tointeger(luaState, 1);
+                        var lookback = argCount == 2 ? lua_toboolean(luaState, 2) : false;");
 
         if (model.IsTag)
             sb.Append(@"
-                        var value = components.HasTagForEntity(entityId);");
+                        var value = components.HasTagForEntity(entityId, lookback);");
         else
             sb.Append(@"
-                        var value = components[entityId];");
+                        var value = lookback ? components.GetComponentWithLookback(entityId) : components[entityId];");
 
         sb.Append($@"